refact(storybook): Move storybook specs
