+++ packages/vlossom/src/components/vs-button/VsButton/stories/VsButton.stories.ts
@@ -1,8 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/vue3';
-
-import VsButton from '../VsButton.vue';
-import { colorScheme } from '@/declaration/storybook/arg-types';
-import { storybookUtil } from '@/utils';
+import { colorScheme, getMetaArguments } from '@/storybook/args';
+import VsButton from './../VsButton.vue';
 
 const meta: Meta<typeof VsButton> = {
     title: 'Components/Base Components/VsButton',
@@ -20,7 +18,7 @@ const meta: Meta<typeof VsButton> = {
     },
 };
 
-meta.args = storybookUtil.getMetaArguments(VsButton.props, meta.args);
+meta.args = getMetaArguments(VsButton.props, meta.args);
 export default meta;
 type Story = StoryObj<typeof VsButton>;
