+++ packages/vlossom/src/components/vs-section/VsSection/stories/VsSection.stories.ts
@@ -1,8 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/vue3';
-
-import VsSection from '../VsSection.vue';
-import { colorScheme } from '@/declaration/storybook/arg-types';
-import { storybookUtil } from '@/utils';
+import { colorScheme, getMetaArguments } from '@/storybook/args';
+import VsSection from './../VsSection.vue';
 
 const meta: Meta<typeof VsSection> = {
     title: 'Components/Layout Components/VsSection',
@@ -20,7 +18,7 @@ const meta: Meta<typeof VsSection> = {
     },
 };
 
-meta.args = storybookUtil.getMetaArguments(VsSection.props, meta.args);
+meta.args = getMetaArguments(VsSection.props, meta.args);
 export default meta;
 type Story = StoryObj<typeof VsSection>;
