--- packages/vlossom/src/declaration/storybook/arg-types.ts
@@ -1,4 +1,0 @@
-export const colorScheme = {
-    control: 'select',
-    options: ['red', 'amber', 'green', 'teal', 'blue', 'indigo', 'purple', 'pink'],
-};
