+++ packages/vlossom/src/storybook/args.ts
@@ -1,0 +1,17 @@
+export const colorScheme = {
+    control: 'select',
+    options: ['red', 'amber', 'green', 'teal', 'blue', 'indigo', 'purple', 'pink'],
+};
+
+export function getMetaArguments(componentProps: { [key: string]: any }, originalArgs: { [key: string]: any } = {}) {
+    const metaArgs: { [key: string]: any } = {};
+    Object.keys(componentProps).forEach((prop) => {
+        const { default: defaultValue } = componentProps[prop];
+        if (typeof defaultValue === 'function') {
+            metaArgs[prop] = defaultValue();
+            return;
+        }
+        metaArgs[prop] = defaultValue ?? null;
+    });
+    return { ...metaArgs, ...originalArgs };
+}
