+++ packages/vlossom/src/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './storybook';
 export * from './string';
